[Refactor] #161 - 루트뷰로 분리
